move saved variant json reload to es specific module
