+++ seqr/management/commands/reload_saved_variant_json.py
@@ -2,6 +2,7 @@ import logging
 from django.core.management.base import BaseCommand, CommandError
 from django.db.models.query_utils import Q
 from seqr.models import Project
+from seqr.utils.search.elasticsearch.es_utils import update_project_saved_variant_json
 from seqr.utils.search.utils import backend_specific_call
 from seqr.views.utils.variant_utils import update_projects_saved_variant_json
 
@@ -33,5 +34,5 @@ class Command(BaseCommand):
 
         family_ids = [family_guid] if family_guid else None
         project_list = [(*project, family_ids) for project in projects.order_by('id').values_list('id', 'guid', 'name', 'genome_version')]
-        update_projects_saved_variant_json(project_list, user_email='manage_command')
+        update_projects_saved_variant_json(project_list, user_email='manage_command', update_function=update_project_saved_variant_json)
         logger.info("Done")
