+++ seqr/utils/search/elasticsearch/es_utils.py
@@ -4,12 +4,13 @@ from elasticsearch.exceptions import ConnectionError as EsConnectionError, Trans
 import elasticsearch_dsl
 from urllib3.connectionpool import connection_from_url
 
-from seqr.models import Sample
+from seqr.models import Sample, SavedVariant
 from seqr.utils.redis_utils import safe_redis_get_json, safe_redis_set_json
 from seqr.utils.search.constants import VCF_FILE_EXTENSIONS, XPOS_SORT_KEY
 from seqr.utils.search.elasticsearch.es_gene_agg_search import EsGeneAggSearch
 from seqr.utils.search.elasticsearch.es_search import EsSearch, get_compound_het_page
 from seqr.views.utils.json_utils import  _to_camel_case
+from seqr.views.utils.variant_utils import get_variants_for_variant_ids, get_saved_variants
 from settings import ELASTICSEARCH_SERVICE_HOSTNAME, ELASTICSEARCH_SERVICE_PORT, ELASTICSEARCH_CREDENTIALS, \
     ELASTICSEARCH_PROTOCOL, ES_SSL_CONTEXT, KIBANA_SERVER
 
@@ -28,6 +29,8 @@ ES_EXCEPTION_MESSAGE_MAP = {
     TransportError: lambda e: '{}: {} - {} - {}'.format(e.__class__.__name__, e.status_code, repr(e.error), _get_transport_error_type(e.info)),
 }
 ES_ERROR_LOG_EXCEPTIONS = {InvalidIndexException}
+
+MAX_VARIANTS_FETCH = 1000
 
 
 def _get_transport_error_type(error):
@@ -330,3 +333,35 @@ def process_es_previously_loaded_gene_aggs(previous_search_results):
             for family_guid in variants[0]['familyGuids']:
                 gene_aggs[gene_id]['families'][family_guid] += len(variants)
     return gene_aggs
+
+
+def update_project_saved_variant_json(project_id, genome_version, family_guids=None, dataset_type=None, user=None, user_email=None, **kwargs):
+    saved_variants = get_saved_variants(genome_version, project_id, family_guids, dataset_type).select_related('family')
+
+    if not saved_variants:
+        return None
+
+    families = set()
+    variant_ids = set()
+    saved_variants_map = {}
+    for v in saved_variants:
+        families.add(v.family)
+        variant_ids.add(v.variant_id)
+        saved_variants_map[(v.variant_id, v.family.guid)] = v
+
+    variant_ids = sorted(variant_ids)
+    families = sorted(families, key=lambda f: f.guid)
+    variants_json = []
+    for sub_var_ids in [variant_ids[i:i+MAX_VARIANTS_FETCH] for i in range(0, len(variant_ids), MAX_VARIANTS_FETCH)]:
+        variants_json += get_variants_for_variant_ids(families, sub_var_ids, user=user, user_email=user_email)
+
+    updated_saved_variants = {}
+    for var in variants_json:
+        for family_guid in var['familyGuids']:
+            saved_variant = saved_variants_map.get((var['variantId'], family_guid))
+            if saved_variant:
+                saved_variant.saved_variant_json = var
+                updated_saved_variants[saved_variant.guid] = saved_variant
+    SavedVariant.bulk_update_models(user, list(updated_saved_variants.values()), ['saved_variant_json'])
+
+    return updated_saved_variants
