+++ seqr/views/apis/saved_variant_api.py
@@ -8,6 +8,7 @@ from clickhouse_search.backend.functions import ArrayFilter, ArrayMap
 from clickhouse_search.search import get_annotations_queryset, get_transcripts_queryset
 from seqr.models import SavedVariant, VariantTagType, VariantTag, VariantNote, VariantFunctionalData,\
     Family, GeneNote, Project
+from seqr.utils.search.elasticsearch.es_utils import update_project_saved_variant_json
 from seqr.utils.search.utils import backend_specific_call
 from seqr.views.utils.json_to_orm_utils import update_model_from_json, get_or_create_model_from_json, \
     create_model_from_json
@@ -16,8 +17,7 @@ from seqr.views.utils.orm_to_json_utils import get_json_for_saved_variants_with_
     get_json_for_saved_variants_child_entities, get_json_for_gene_notes_by_gene_id, STRUCTURED_METADATA_TAG_TYPES
 from seqr.views.utils.permissions_utils import get_project_and_check_permissions, check_project_permissions, \
     login_and_policies_required
-from seqr.views.utils.variant_utils import update_project_saved_variant_json, reset_cached_search_results, \
-    get_variants_response, parse_saved_variant_json
+from seqr.views.utils.variant_utils import reset_cached_search_results, get_variants_response, parse_saved_variant_json
 
 
 logger = logging.getLogger(__name__)
